requirements.txt
agent-framework
python-dotenv
pydantic